Recognize blog posts by tag Post instead of TODO DONE state
This commit changes the behavior of the Org-Mode backend of o-blog
such that blog posts are no longer identified by the TODO state DONE.
Instead the backend will now find sections that have the tag :Post:
and export those ones. The tag :Post: will not be considered for the
tag cloud.

This change breaks prior behavior and uses a hard coded tag to
identifiy blog posts. It would be better to make this configurable
such that the user can choose wether blog posts are identified by
a certain TODO state, or CLOSED timestamp or a list of tags.